Move popup tests to skikoTest
