+++ compose/ui/ui/src/desktopTest/kotlin/androidx/compose/ui/window/DesktopPopupTest.kt
@@ -22,110 +22,17 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.runtime.*
 import androidx.compose.runtime.snapshots.Snapshot
-import androidx.compose.ui.*
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.input.pointer.PointerEventType
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.Layout
-import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.test.junit4.createComposeRule
-import androidx.compose.ui.unit.Density
-import androidx.compose.ui.unit.IntRect
-import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
 import com.google.common.truth.Truth.assertThat
-import kotlin.test.fail
 import org.junit.Rule
 import org.junit.Test
 
-// TODO: Move in common once Popup is available there
 class DesktopPopupTest {
     @get:Rule
     val rule = createComposeRule()
-
-    @Test
-    fun `pass composition locals to popup`() {
-        val compositionLocal = staticCompositionLocalOf<Int> {
-            error("not set")
-        }
-
-        var actualLocalValue = 0
-
-        rule.setContent {
-            CompositionLocalProvider(compositionLocal provides 3) {
-                Popup {
-                    actualLocalValue = compositionLocal.current
-                }
-            }
-        }
-
-        assertThat(actualLocalValue).isEqualTo(3)
-    }
-
-    // https://github.com/JetBrains/compose-multiplatform/issues/3142
-    @Test
-    fun `pass LayoutDirection to popup`() {
-        lateinit var localLayoutDirection: LayoutDirection
-
-        var layoutDirection by mutableStateOf(LayoutDirection.Rtl)
-        rule.setContent {
-            CompositionLocalProvider(LocalLayoutDirection provides layoutDirection) {
-                Popup {
-                    localLayoutDirection = LocalLayoutDirection.current
-                }
-            }
-        }
-
-        assertThat(localLayoutDirection).isEqualTo(LayoutDirection.Rtl)
-
-        // Test that changing the local propagates it into the popup
-        layoutDirection = LayoutDirection.Ltr
-        rule.waitForIdle()
-        assertThat(localLayoutDirection).isEqualTo(LayoutDirection.Ltr)
-    }
-
-    @Test
-    fun `onDispose inside popup`() {
-        var isPopupShowing by mutableStateOf(true)
-        var isDisposed = false
-
-        rule.setContent {
-            if (isPopupShowing) {
-                Popup {
-                    DisposableEffect(Unit) {
-                        onDispose {
-                            isDisposed = true
-                        }
-                    }
-                }
-            }
-        }
-
-        isPopupShowing = false
-        rule.waitForIdle()
-
-        assertThat(isDisposed).isEqualTo(true)
-    }
-
-    @Test
-    fun `use density inside popup`() {
-        var density by mutableStateOf(Density(2f, 1f))
-        var densityInsidePopup = 0f
-
-        rule.setContent {
-            CompositionLocalProvider(LocalDensity provides density) {
-                Popup {
-                    densityInsidePopup = LocalDensity.current.density
-                }
-            }
-        }
-
-        assertThat(densityInsidePopup).isEqualTo(2f)
-
-        density = Density(3f, 1f)
-        rule.waitForIdle()
-        assertThat(densityInsidePopup).isEqualTo(3f)
-    }
 
     @Test(timeout = 5000) // TODO(demin): why, when an error has occurred, this test never ends?
     fun `(Bug) after open popup use derivedStateOf inside main window draw`() {
@@ -239,257 +146,5 @@ class DesktopPopupTest {
         }
 
         rule.waitForIdle()
-    }
-
-    @Test
-    fun `call dismiss if clicked outside of focusable popup`() = ImageComposeScene(
-        100,
-        100
-    ).use { scene ->
-        var onDismissRequestCallCount = 0
-
-        val background = FillBox()
-        val popup = PopupState(
-            IntRect(20, 20, 60, 60),
-            focusable = true,
-            onDismissRequest = { onDismissRequestCallCount++ }
-        )
-
-        scene.setContent {
-            background.Content()
-            popup.Content()
-        }
-
-        assertThat(onDismissRequestCallCount).isEqualTo(0)
-        scene.sendPointerEvent(PointerEventType.Press, Offset(10f, 10f))
-        background.events.assertReceivedNoEvents()
-        assertThat(onDismissRequestCallCount).isEqualTo(1)
-    }
-
-    @Test
-    fun `call dismiss if clicked outside of non-focusable popup`() = ImageComposeScene(
-        100,
-        100
-    ).use { scene ->
-        var onDismissRequestCallCount = 0
-
-        val background = FillBox()
-        val popup = PopupState(
-            IntRect(20, 20, 60, 60),
-            focusable = false,
-            dismissOnClickOutside = true,
-            onDismissRequest = { onDismissRequestCallCount++ }
-        )
-
-        scene.setContent {
-            background.Content()
-            popup.Content()
-        }
-
-        assertThat(onDismissRequestCallCount).isEqualTo(0)
-        scene.sendPointerEvent(PointerEventType.Press, Offset(10f, 10f))
-        background.events.assertReceivedLast(PointerEventType.Press, Offset(10f, 10f))
-        assertThat(onDismissRequestCallCount).isEqualTo(1)
-    }
-
-    @Test
-    fun `call dismiss if clicked outside of multiple non-focusable popups`() = ImageComposeScene(
-        100,
-        100
-    ).use { scene ->
-        var onDismissRequestCallCount = 0
-
-        val background = FillBox()
-        val popup1 = PopupState(
-            IntRect(20, 20, 60, 60),
-            focusable = false,
-            dismissOnClickOutside = true,
-            onDismissRequest = { onDismissRequestCallCount++ }
-        )
-        val popup2 = PopupState(
-            IntRect(30, 30, 70, 70),
-            focusable = false,
-            dismissOnClickOutside = true,
-            onDismissRequest = { onDismissRequestCallCount++ }
-        )
-
-        scene.setContent {
-            background.Content()
-            popup1.Content()
-            popup2.Content()
-        }
-
-        assertThat(onDismissRequestCallCount).isEqualTo(0)
-        scene.sendPointerEvent(PointerEventType.Press, Offset(10f, 10f))
-        background.events.assertReceivedLast(PointerEventType.Press, Offset(10f, 10f))
-        assertThat(onDismissRequestCallCount).isEqualTo(2)
-    }
-
-    @Test
-    fun `call dismiss for non-focusable popups above`() = ImageComposeScene(
-        100,
-        100
-    ).use { scene ->
-        var onDismissRequestCallCount = 0
-
-        val background = FillBox()
-        val popup1 = PopupState(
-            IntRect(10, 10, 50, 50),
-            focusable = false,
-            dismissOnClickOutside = true,
-            onDismissRequest = { fail() }
-        )
-        val popup2 = PopupState(
-            IntRect(20, 20, 60, 60),
-            focusable = false,
-            dismissOnClickOutside = true,
-            onDismissRequest = { fail() }
-        )
-        val popup3 = PopupState(
-            IntRect(30, 30, 70, 70),
-            focusable = false,
-            dismissOnClickOutside = true,
-            onDismissRequest = { onDismissRequestCallCount++ }
-        )
-        val popup4 = PopupState(
-            IntRect(40, 40, 80, 80),
-            focusable = false,
-            dismissOnClickOutside = true,
-            onDismissRequest = { onDismissRequestCallCount++ }
-        )
-
-        scene.setContent {
-            background.Content()
-            popup1.Content()
-            popup2.Content()
-            popup3.Content()
-            popup4.Content()
-        }
-
-        assertThat(onDismissRequestCallCount).isEqualTo(0)
-        scene.sendPointerEvent(PointerEventType.Press, Offset(55f, 25f))
-        background.events.assertReceivedNoEvents()
-        assertThat(onDismissRequestCallCount).isEqualTo(2)
-    }
-
-    @Test
-    fun `call dismiss for above focusable popup`() = ImageComposeScene(
-        100,
-        100
-    ).use { scene ->
-        var onDismissRequestCallCount = 0
-
-        val background = FillBox()
-        val popup1 = PopupState(
-            IntRect(10, 10, 50, 50),
-            focusable = false,
-            dismissOnClickOutside = true,
-            onDismissRequest = { fail() }
-        )
-        val popup2 = PopupState(
-            IntRect(20, 20, 60, 60),
-            focusable = false,
-            dismissOnClickOutside = true,
-            onDismissRequest = { fail() }
-        )
-        val popup3 = PopupState(
-            IntRect(30, 30, 70, 70),
-            focusable = true,
-            dismissOnClickOutside = true,
-            onDismissRequest = { onDismissRequestCallCount++ }
-        )
-        val popup4 = PopupState(
-            IntRect(40, 40, 80, 80),
-            focusable = false,
-            dismissOnClickOutside = true,
-            onDismissRequest = { onDismissRequestCallCount++ }
-        )
-
-        scene.setContent {
-            background.Content()
-            popup1.Content()
-            popup2.Content()
-            popup3.Content()
-            popup4.Content()
-        }
-
-        assertThat(onDismissRequestCallCount).isEqualTo(0)
-        scene.sendPointerEvent(PointerEventType.Press, Offset(5f, 5f))
-        background.events.assertReceivedNoEvents()
-        assertThat(onDismissRequestCallCount).isEqualTo(2)
-    }
-
-    @Test
-    fun `pass event if clicked outside of non-focusable popup`() = ImageComposeScene(
-        100,
-        100
-    ).use { scene ->
-        var onDismissRequestCallCount = 0
-
-        val background = FillBox()
-        val popup = PopupState(
-            IntRect(20, 20, 60, 60),
-            focusable = false,
-            onDismissRequest = { onDismissRequestCallCount++ }
-        )
-
-        scene.setContent {
-            background.Content()
-            popup.Content()
-        }
-
-        assertThat(onDismissRequestCallCount).isEqualTo(0)
-        scene.sendPointerEvent(PointerEventType.Press, Offset(10f, 10f))
-        background.events.assertReceivedLast(PointerEventType.Press, Offset(10f, 10f))
-        assertThat(onDismissRequestCallCount).isEqualTo(0)
-    }
-
-    @Test
-    fun `don't pass event if clicked outside of focusable popup`() = ImageComposeScene(
-        100,
-        100
-    ).use { scene ->
-        val background = FillBox()
-        val popup = PopupState(
-            IntRect(20, 20, 60, 60),
-            focusable = true
-        )
-
-        scene.setContent {
-            background.Content()
-            popup.Content()
-        }
-
-        scene.sendPointerEvent(PointerEventType.Press, Offset(10f, 10f))
-        scene.sendPointerEvent(PointerEventType.Release, Offset(10f, 10f))
-        background.events.assertReceivedNoEvents()
-    }
-
-    @Test
-    fun `can scroll outside of non-focusable popup`() = ImageComposeScene(100, 100).use { scene ->
-        val background = FillBox()
-        val popup = PopupState(IntRect(20, 20, 60, 60), focusable = false)
-
-        scene.setContent {
-            background.Content()
-            popup.Content()
-        }
-
-        scene.sendPointerEvent(PointerEventType.Scroll, Offset(10f, 10f))
-        background.events.assertReceivedLast(PointerEventType.Scroll, Offset(10f, 10f))
-    }
-
-    @Test
-    fun `can't scroll outside of focusable popup`() = ImageComposeScene(100, 100).use { scene ->
-        val background = FillBox()
-        val popup = PopupState(IntRect(20, 20, 60, 60), focusable = true)
-
-        scene.setContent {
-            background.Content()
-            popup.Content()
-        }
-
-        scene.sendPointerEvent(PointerEventType.Scroll, Offset(10f, 10f))
-        background.events.assertReceivedNoEvents()
     }
 }
